+++ Changelog.md
@@ -7,6 +7,7 @@
 - Fix order of "last ten runs" on the flows page [#50](https://github.com/PrefectHQ/ui/pull/50)
 - Fix issue with Failed Tasks tile incorrectly displaying with a red color - [#88](https://github.com/PrefectHQ/ui/issues/88)
 - Add cancellation button to Flow Run pages [#92](https://github.com/PrefectHQ/ui/pull/92)
+- Add text-area to allow a user to add and/or remove labels from the flow details tile [#106](https://github.com/PrefectHQ/ui/pull/106)
 - Name the restart mutation and fix casing on the restart mutation variable [#98](https://github.com/PrefectHQ/ui/pull/98)
 - Add Created On column to Flow Table - [#103](https://github.com/PrefectHQ/ui/pull/103)
 - Improve schedule toggling reactivity [#104](https://github.com/PrefectHQ/ui/pull/104/files)